Updates to scenes before early release of (the actual) beta2
